refactor: Refactor backend structure
