+++ back-end/api/__init__.py
@@ -1,4 +1,5 @@
-from . import api
-from . import predictor 
-from . import s3
-from . import uploader 
+from . import config
+from . import controller
+from . import service
+from . import common 
+from . import model
