+++ back-end/api/service/user_statistic.py
@@ -1,0 +1,5 @@
+from sqlalchemy.orm import Session
+from ..model import models, schemas
+
+def get_user_statistic(db: Session):
+    return db.query(UserStatistic).all()
